Refactor all modal components to one modal component
