+++ src/components/rules/UpdateRuleForm.tsx
@@ -1,14 +1,18 @@
-import { useAttributes } from "@/hooks/attributes";
 import { useForm } from "react-hook-form";
 import { useDeleteSegmentRule, useUpdateSegmentRule } from "@/hooks/segments";
 import FormButton from "@/components/ui/FormButton";
-import { operators } from "@/utils/data";
+import {
+  booleanOperators,
+  numberOperators,
+  stringOperators,
+} from "@/utils/data";
 import toast from "react-hot-toast";
 import ToastTUI from "../ToastTUI";
 import { AxiosError } from "axios";
 import FormHeader from "../ui/FormHeader";
-import { useState } from "react";
-import DeleteModal from "../DeleteModal";
+import { RuleFormInputs } from "./CreateRuleForm";
+import { Attribute } from "@/types";
+import FormInput from "../ui/FormInput";
 
 type UpdateRuleFormProps = {
   sKey: string;
@@ -16,6 +20,7 @@ type UpdateRuleFormProps = {
   aKey: string;
   operator: string;
   value: string;
+  attributes: Attribute[];
   setOpen: React.Dispatch<React.SetStateAction<boolean>>;
 };
 
@@ -26,22 +31,65 @@ function UpdateRuleForm({
   operator,
   value,
   setOpen,
+  attributes,
 }: UpdateRuleFormProps) {
-  const { data: attributes } = useAttributes();
-  const { register, handleSubmit } = useForm<{
-    attribute: string;
-    operator: string;
-    value: string;
-    sKey: string;
-  }>();
+  const {
+    register,
+    handleSubmit,
+    watch,
+    setError,
+    formState: { errors },
+  } = useForm<RuleFormInputs>({
+    defaultValues: {
+      attribute: attributes.find((a) => a.aKey === aKey)!.name,
+      value,
+      sKey,
+    },
+  });
 
   const { mutateAsync: updateSegmentRuleMutate } = useUpdateSegmentRule();
 
+  const currDataType = attributes.find(
+    (a) => a.name === watch("attribute"),
+  )!.type;
+
+  const operators = (() => {
+    switch (currDataType) {
+      case "string":
+        return stringOperators;
+      case "number":
+        return numberOperators;
+      case "boolean":
+        return booleanOperators;
+    }
+  })();
+
   const onSubmit = handleSubmit(async ({ attribute, operator, value }) => {
-    if (!attributes) return;
-    const attr = attributes.find((a) => a.name === attribute)!.aKey;
+    const attr = attributes.find((a) => a.name === attribute)!;
+
+    switch (attr.type) {
+      case "boolean":
+        if (value !== "true" && value !== "false") {
+          setError("value", { message: "Must be true or false" });
+          return;
+        }
+        break;
+      case "number":
+        if (!value || Number.isNaN(Number(value))) {
+          setError("value", { message: "The value must be a number" });
+          return;
+        }
+        break;
+      case "string":
+        if (!value) {
+          setError("value", { message: "Value is required" });
+          return;
+        }
+        break;
+    }
+
     const data = {
-      aKey: attr,
+      aKey: attr.aKey,
       operator,
       value,
       sKey,
@@ -62,7 +110,7 @@ function UpdateRuleForm({
     setOpen(false);
   });
 
-  const [openDeleteModal, setOpenDeleteModal] = useState(false);
+  // const [openDeleteModal, setOpenDeleteModal] = useState(false);
   const { mutateAsync: deleteRuleMutate } = useDeleteSegmentRule();
   const handleDelete = async () => {
     try {
@@ -76,7 +124,7 @@ function UpdateRuleForm({
         toast.custom(<ToastTUI type="error" message={responseError} />);
       }
     }
-    setOpenDeleteModal(false);
+    // setOpenDeleteModal(false);
     setOpen(false);
   };
 
@@ -103,15 +151,12 @@ function UpdateRuleForm({
             <select
               id="attribute"
               className="mt-2 block w-full rounded-md border-0 py-1.5 pl-3 pr-10 text-gray-900 ring-1 ring-inset ring-gray-300 focus:ring-2 focus:ring-ss-blgr sm:text-sm sm:leading-6"
-              defaultValue={attributes?.find((a) => a.aKey === aKey)!.name}
               {...register("attribute")}
             >
-              {attributes?.length === 0 ? (
+              {attributes.length === 0 ? (
                 <option disabled>No attributes</option>
               ) : (
-                attributes?.map(({ name }) => (
-                  <option key={name}>{name}</option>
-                ))
+                attributes.map(({ name }) => <option key={name}>{name}</option>)
               )}
             </select>
           </div>
@@ -141,13 +186,12 @@ function UpdateRuleForm({
               Value
             </label>
             <div className="mt-2">
-              <input
-                type="text"
+              <FormInput
                 id="value"
-                {...register("value")}
-                className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-ss-blgr sm:text-sm sm:leading-6"
-                defaultValue={value}
                 placeholder="Enter a value"
+                register={register("value")}
+                isError={!!errors.value}
+                errorMessage={errors.value?.message}
               />
             </div>
           </div>
@@ -162,12 +206,6 @@ function UpdateRuleForm({
           <FormButton typeOfButton="confirm" type="submit" text="Save" />
         </div>
       </form>
-      <DeleteModal
-        setOpen={setOpenDeleteModal}
-        open={openDeleteModal}
-        resource="rule"
-        onDelete={handleDelete}
-      />
     </>
   );
 }
